requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.26.0
python-dotenv==0.19.0
feedgen==1.0.0